just trying to get rid of these large files

+++ aya_scraper.py
@@ -12,15 +12,16 @@ from selenium.webdriver.support.ui import WebDriverWait
 from selenium.webdriver.support import expected_conditions as EC
 from selenium.common.exceptions import NoSuchElementException
 from selenium.common.exceptions import ElementClickInterceptedException
+from selenium.common.exceptions import TimeoutException
 import time
 
 # modify parameters 
 STATE = "California"
+SPECIALTIES = ["Telemetry", "Tele Float", "Medical Surgical", "Medical Surgical Float", "Medical Surgical Oncology", "MS/Tele", "MS/Tele Float"]
 
 #folder in which to save the html files inside of the current directory
 folder_path = "./aya"
 
-#TODO add functionalty to select individual specialties
 
 def set_up_driver():
     from selenium.webdriver.chrome.options import Options
@@ -29,6 +30,7 @@ def set_up_driver():
 
     # Setup Chrome options
     chrome_options = Options()
+    #chrome_options.add_argument("--headless")  # Ensure GUI is off
     chrome_options.add_argument("--disable-extensions")
     chrome_options.add_argument("--disable-gpu")
     chrome_options.add_argument("--no-sandbox") # Bypass OS security model
@@ -80,7 +82,7 @@ def select_location(driver, STATE):
     # Click the "search now" button, which is second of two items on this list
     search_buttons[1].click()
 
-
+    
 # Select any or all nursing specialties from dropdown
 def select_specialty(driver):
     
@@ -106,10 +108,17 @@ def select_specialty(driver):
     # Wait for third dropdown 
     wait.until(EC.visibility_of_element_located((By.CLASS_NAME, "options--expertises")))
 
-    #TODO add functionalty to select individual specialties
-    # Click 'all specialties' in the third dropdown
-    expertise = driver.find_element(By.CLASS_NAME, "expertise--all")
-    expertise.click()
+    # Select specific specialties
+    for specialty in SPECIALTIES:
+        try:
+            specialty_element = driver.find_element(By.XPATH, f'//a[contains(@class, "expertise") and @data-expertisetext="{specialty}"]')
+            specialty_element.click()
+        except NoSuchElementException:
+            print(f"Specialty '{specialty}' not found.")
+
+    # # Click 'all specialties' in the third dropdown
+    # expertise = driver.find_element(By.CLASS_NAME, "expertise--all")
+    # expertise.click()
 
     # wait for the search button to become clickable
     wait = WebDriverWait(driver, 10)
@@ -123,8 +132,32 @@ def select_specialty(driver):
     search_button.click()
 
 
+def click_next_button(driver, next_button):
+    try:
+        wait = WebDriverWait(driver, 3)  # Increase the timeout value to 30 seconds
+        wait.until_not(EC.presence_of_element_located((By.CSS_SELECTOR, "button#nextButton.is-loading")))
+        next_button = wait.until(EC.element_to_be_clickable((By.XPATH, "//button[@id='nextButton']")))
+
+        next_button.click()
+
+    except (TimeoutException, ElementClickInterceptedException):
+        print("Timeout or click interception occurred. Retrying with scrolling...")
+        
+        # Scroll to the bottom of the page
+        driver.execute_script("window.scrollTo(0, document.body.scrollHeight);")
+        time.sleep(1)
+        # scroll back up.
+        x = next_button.location["x"]
+        y = next_button.location["y"]
+        driver.execute_script(f"window.scrollTo({x}, {y - 500})")
+        # wait a tic
+        time.sleep(1)
+        
+        # try clicking again
+        next_button.click()
+
+
 def navigate_pages(driver):
-    
     # Scrape the first page
     html_content = ""
 
@@ -156,24 +189,8 @@ def navigate_pages(driver):
             y = next_button.location["y"]
             driver.execute_script(f"window.scrollTo({x}, {y - 500})")
 
-            # Click the next button using WebDriverWait
-            try:
-                time.sleep(1)
-                wait = WebDriverWait(driver, 10)  # Adjust the timeout value as needed (10 seconds here)
-                wait.until_not(EC.presence_of_element_located((By.CSS_SELECTOR, "button#nextButton.is-loading")))
-                next_button = wait.until(EC.element_to_be_clickable((By.XPATH, "//button[@id='nextButton']")))
-                next_button.click()
-            except ElementClickInterceptedException:
-                print("Had to scroll again to click element.")
-                
-                # Try to scroll to button again, wait a little longer this time.
-                x = next_button.location["x"]
-                y = next_button.location["y"]
-                driver.execute_script(f"window.scrollTo({x}, {y - 500})")
-                time.sleep(3)
-                
-                # try clicking again
-                next_button.click()
+            # Click the next button
+            click_next_button(driver, next_button)
 
             time.sleep(1)  # wait for the page to load
             page_count += 1
@@ -187,21 +204,37 @@ def navigate_pages(driver):
 
     return html_content
 
-
 def save_html_content(html_content, directory, file_name):
     # Create the directory if it doesn't exist
     os.makedirs(directory, exist_ok=True)
 
     # Save the HTML content to a file in the specified directory
     file_path = os.path.join(directory, file_name)
+    
+    # Check if a file with the same name already exists
+    if os.path.isfile(file_path):
+        print(f"A file named {file_name} already exists in the directory {directory}.")
+        return
+    
     with open(file_path, "w") as f:
         f.write(html_content)
     print(f"Data loaded to {file_path}")
 
+
+def safe_click(driver, by, value):
+    try:
+        element = WebDriverWait(driver, 10).until(EC.element_to_be_clickable((by, value)))
+        element.click()
+    except TimeoutException:
+        print("Element is not clickable yet, waiting for 5 more seconds.")
+        time.sleep(5)
+        safe_click(driver, by, value)
+
+
 def main():
     driver = set_up_driver()
     load_website(driver)
-    #select_location(driver, STATE)
+    # select_location(driver, STATE)
     select_specialty(driver)
     navigate_pages(driver)
     driver.quit()
